+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Changed
+- Rename `getQueueAvailability` to `getAgentAvailability`
 
 ## [0.3.0] - 2021-09-23
 ### Added